server/requirements.txt
# fastapi>=0.104.1
# uvicorn[standard]>=0.24.0
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.7
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
# python-multipart>=0.0.6
# pydantic>=2.4.0
# python-dotenv>=1.0.0
# email-validator>=2.0.0

fastapi
uvicorn
langchain
chromadb
sentence-transformers
transformers
accelerate
torch
python-magic

unstructured
pinecone-client
pypdf
python-docx
faiss-cpu
pdfminer.six
pytesseract
pillow
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
email-validator
